Added post editing, post deletion, added aws s3 file deletion function

+++ CopyCube_Backend/src/post/post.service.ts
@@ -51,6 +51,9 @@ export class PostService {
     if (!post) {
       throw new Error('Post not found');
     }
+    if (post.user.id !== user_id){
+      throw new Error('This post does not belong to you');
+    }
 
     const params = {
         key:hash_id,
@@ -64,11 +67,15 @@ export class PostService {
     await this.postsRepository.save(post);
     return new HttpException("Changes applied successfully", HttpStatus.OK)
   }
-  async remove(hash_id: string){
+  async remove(hash_id: string, user_id: number){
     const post = await this.postsRepository.findOne({where:{key:hash_id}})
     if (!post){
       throw new Error('Post not found');
     }
+    if (post.user.id !== user_id){
+      throw new Error('This post does not belong to you');
+    }
+
     await this.filesService.removeFile(hash_id, this.configService.get('BUCKET_POSTS'))
     await this.postsRepository.remove(post)
